// frontend/src/lib/api/models/SmallGroupLeaderInputDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SmallGroupLeaderInputDto = {
    /**
     * UUID del usuario líder (con login). Uno de los dos debe estar presente.
     */
    leaderUserId?: string | null;
    /**
     * UUID de la persona líder (sin login). Uno de los dos debe estar presente.
     */
    leaderPersonId?: string | null;
};